backend/api/pagination.py
from django.conf import settings

from rest_framework.pagination import PageNumberPagination


class PageLimitPagination(PageNumberPagination):
    page_size = settings.DEFAULT_PAGINATION_PAGE_SIZE
    page_size_query_param = "limit"
