backend/requirements.txt
Django==3.2
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.0.0
drf-extra-fields==3.5.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
python-dotenv==1.0.0
gunicorn==20.1.0
